final code cleanup

+++ TrackEditor/TrackEditor/Curves/Curve.cpp
@@ -1,17 +1,10 @@
-//
-//  Curve.cpp
-//  TrackEditor
-//
-//  Created by Matheus Polonia on 22/11/23.
-//
-
 #include "Curve.hpp"
 
 void Curve::drawCurve(Shader* shader) {
     shader->use();
 
     glBindVertexArray(VAO);
-    glDrawArrays(GL_LINE_LOOP, 0, curvePoints.size());
+    glDrawArrays(GL_LINE_LOOP, 0, (int)curvePoints.size());
     glBindVertexArray(0);
 }
